Abstract the <UserInput> into a <FetchInput>, for reuse
As Luke suggested,
this can allow reusing it for categories and tags.
It's also possible to reuse it for posts,
but this might have to store both the
post ID, which is immutable, and the post title.
The user probably won't want to see the post ID.
Also, the URLInput from Gutenberg stores
the post object and the URL.