Add a method to filter the control's output value
Ideally, this logic would be in each control class.
For example, the user logic would be in the User class.
But it looks like controls are only instantiated in /wp-admin,
as register_controls() runs on 'admin_init'.
Adding a filter might not be the best approach.
We're of course committing to this approach,
as this will be a public API once we release this.