Revert bfde7, store the displayValue in an attribute
It took too long to get the post title from the apiFetch
whenever the Post control mounted.
There probably needs to be a way to store the title (displayValue).
This doesn't seem ideal, but it removes the need to
call apiFetch to get the post title,
or the Taxonomy name for a future Taxonomy control.