If the pro license expires, don't overwrite previous values
As Luke mentioned, this allows someone to
renew the pro license without losing the values.
This will disable and make readonly the fields.
readonly is useful because it allows the name
to be part of the POST request to update the block.
A disabled <input> isn't part of the POST request.