Address an issue where fields from other blocks appeared in the Inspector
This looks to fix one case of the issue,
but I'm not sure if it will fix all cases.
The issue didnt' always appear, even without this.
The field was actually an object,
and it was attempting to concatenate it with a string.
So this is probably something that should be addressed either way.