Like in Luke's commit, throw an Exception if the method doesn't exist
In the __call() method for the Plugin,
I didn't have any handling for the method not existing.
This masked a problem, where Loader
called block_lab->located_template()
before the Utils object was available.
This didn't throw an Exception,
and failed silently.
So this moves the logic that adds Utils to Plugin earlier,
making it available.
It also throws an exception on calling something like
block_lab->nonexistent_method()