Prevent repeated calls to getblocklab.com if the validation request fails
Before, if the wp_remote_post() call is_wp_error(),
like if it had a 404,
it wouldn't save that in the transient.
So it would repeatedly make requests to getblocklab.com,
even multiple times per pageload.
This sets a transient, so that once there's a request that fails,
it won't check again for 24 hours,
just like if a request returned a 'license' of 'invalid'.
This should limit the amount of requests.