Add a workaround for empty rows not appearing in the Repeater
Before, on adding empty rows,
they did not appear at all in the editor.
This looks to be from the way <ServerSideRender>
handles empty {}.
So this adds a workaround for that, setting
the new row to { '': '' }.
This still isn't ideal, and I should keep an eye
out for a better approach.