Add a function to reset rows, thanks to Ben Harris
As reported by Ben and Markus,
it wasn't possible to have 2 of the
same repeater block on a post.
So this function simply calls:
block_lab()->loop()->reset( $name ),
on Ben's suggestion.
It acts like wp_reset_postdata(),
and should go at the end of the block that begins with:
if ( block_rows(  ) ) :